langchain_chatbot/requirements.txt
flask
streamlit
openai
faiss-cpu
langchain
sentence-transformers
python-dotenv
requests
